Avoid unwanted float to int conversion in ScoreEndPosition() introduced by recent changes
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@260 2e953b5c-c64d-0410-be54-f773e93e544c

+++ go/GoBoardUtil.h
@@ -471,7 +471,7 @@ float GoBoardUtil::ScoreEndPosition(const BOARD& bd, float komi,
                                     const SgBWSet& safe, bool noCheck,
                                   SgPointArray<SgEmptyBlackWhite>* scoreBoard)
 {
-    int score = -komi;
+    float score = -komi;
     for (typename BOARD::Iterator it(bd); it; ++it)
     {
         SgPoint p = *it;
